refactor(abw): decompose day blurb composition
